+++ include/bitcoin/bitcoin/message/block.hpp
@@ -70,7 +70,7 @@ public:
     data_chunk to_data(uint32_t version) const;
     void to_data(uint32_t version, std::ostream& stream) const;
     void to_data(uint32_t version, writer& sink) const;
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     block& operator=(chain::block&& other);
