+++ include/bitcoin/bitcoin/message/filter_load.hpp
@@ -74,7 +74,7 @@ public:
     void to_data(uint32_t version, writer& sink) const;
     bool is_valid() const;
     void reset();
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     // This class is move assignable but not copy assignable.
     filter_load& operator=(filter_load&& other);
