+++ include/bitcoin/bitcoin/message/transaction.hpp
@@ -75,7 +75,7 @@ public:
     data_chunk to_data(uint32_t version=version::level::canonical) const;
     void to_data(uint32_t version, std::ostream& stream) const;
     void to_data(uint32_t version, writer& sink) const;
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     transaction& operator=(chain::transaction&& other);
