+++ include/bitcoin/bitcoin/message/verack.hpp
@@ -41,7 +41,7 @@ public:
     static verack factory_from_data(uint32_t version, const data_chunk& data);
     static verack factory_from_data(uint32_t version, std::istream& stream);
     static verack factory_from_data(uint32_t version, reader& source);
-    static uint64_t satoshi_fixed_size(uint32_t version);
+    static size_t satoshi_fixed_size(uint32_t version);
 
     verack();
 
@@ -53,7 +53,7 @@ public:
     void to_data(uint32_t version, writer& sink) const;
     bool is_valid() const;
     void reset();
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     static const std::string command;
     static const uint32_t version_minimum;
