+++ src/message/inventory_vector.cpp
@@ -179,12 +179,12 @@ void inventory_vector::to_data(uint32_t version,
     sink.write_hash(hash_);
 }
 
-uint64_t inventory_vector::serialized_size(uint32_t version) const
+size_t inventory_vector::serialized_size(uint32_t version) const
 {
     return inventory_vector::satoshi_fixed_size(version);
 }
 
-uint64_t inventory_vector::satoshi_fixed_size(uint32_t version)
+size_t inventory_vector::satoshi_fixed_size(uint32_t version)
 {
     return sizeof(hash_) + sizeof(uint32_t);
 }
