+++ src/message/memory_pool.cpp
@@ -139,12 +139,12 @@ void memory_pool::to_data(uint32_t version, writer& sink) const
 {
 }
 
-uint64_t memory_pool::serialized_size(uint32_t version) const
+size_t memory_pool::serialized_size(uint32_t version) const
 {
     return memory_pool::satoshi_fixed_size(version);
 }
 
-uint64_t memory_pool::satoshi_fixed_size(uint32_t version)
+size_t memory_pool::satoshi_fixed_size(uint32_t version)
 {
     return 0;
 }
